+++ changelog.md
@@ -15,6 +15,7 @@
 * Resolved issues with having only a header image, a header image with a logo, a header image with text, and a logo or text without a header image
 * General SEO improvements to the header
 * Resolving a problem that prevented the Personal Image widget placeholder from displaying in Firefox
+* Introducing the ability to reset the social icons (ideal for those who ran previews, betas, or who want to clear out their available icons)
 
 ### Influence Widget